+++ package.json
@@ -19,16 +19,15 @@
     "@mui/icons-material": "7.3.6",
     "@mui/material": "7.3.6",
     "@mui/material-nextjs": "7.3.6",
-    "next": "^16.1.1",
+    "next": "16.1.1",
     "react": "18",
     "react-dom": "18",
     "sass": "1.77.8"
   },
   "devDependencies": {
-    "@chromatic-com/storybook": "1.6.1",
-    "@storybook/addon-links": "9.1.17",
-    "@storybook/addon-onboarding": "9.1.17",
-    "@storybook/nextjs": "9.1.17",
+    "@chromatic-com/storybook": "4.1.3",
+    "@storybook/addon-links": "10.1.10",
+    "@storybook/addon-onboarding": "10.1.10",
     "@types/node": "20",
     "@types/react": "18",
     "@types/react-dom": "18",
@@ -36,16 +35,18 @@
     "eslint-config-next": "14.2.3",
     "eslint-config-prettier": "9.1.0",
     "eslint-plugin-prettier": "5.2.1",
-    "eslint-plugin-storybook": "9.1.17",
+    "eslint-plugin-storybook": "10.1.10",
     "postcss": "8",
     "postcss-styled-syntax": "0.6.4",
     "prettier": "3.3.3",
-    "storybook": "9.1.17",
+    "storybook": "10.1.10",
     "stylelint": "16.7.0",
     "stylelint-config-standard": "36.0.1",
     "stylelint-order": "6.0.4",
     "tailwindcss": "3.4.1",
     "typescript": "5",
-    "@storybook/addon-docs": "9.1.17"
+    "@storybook/addon-docs": "10.1.10",
+    "@storybook/nextjs-vite": "10.1.10",
+    "vite": "7.0.0"
   }
 }
