+++ src/components/parts/Card/index.stories.tsx
@@ -1,6 +1,5 @@
-import { Layout } from "../Layout";
 import { Card } from ".";
-import type { Meta, StoryObj } from "@storybook/nextjs";
+import type { Meta, StoryObj } from "@storybook/nextjs-vite";
 
 const meta: Meta<typeof Card> = {
   component: Card,
@@ -9,19 +8,10 @@ const meta: Meta<typeof Card> = {
 export default meta;
 type Story = StoryObj<typeof Card>;
 
-const Component: Story["render"] = (args) => {
-  return (
-    <Layout>
-      <Card {...args} />
-    </Layout>
-  );
-};
-
 export const Default: Story = {
   args: {
     title: "TypeScript",
     children: "Experienced most of the developments with this.",
     avatarName: "ts.png",
   },
-  render: Component,
 };
