+++ src/components/parts/DrawerNav/index.stories.tsx
@@ -1,8 +1,7 @@
 import { Button } from "@mui/material";
 import { useState } from "react";
-import { Layout } from "../Layout";
 import { DrawerNav } from ".";
-import type { Meta, StoryObj } from "@storybook/nextjs";
+import type { Meta, StoryObj } from "@storybook/nextjs-vite";
 
 const meta: Meta<typeof DrawerNav> = {
   component: DrawerNav,
@@ -15,14 +14,14 @@ const Component: Story["render"] = () => {
   const [isOpen, setIsOpen] = useState(false);
 
   return (
-    <Layout>
+    <>
       <Button onClick={() => setIsOpen(true)}>drawer open</Button>
       <DrawerNav
         isOpen={isOpen}
         onClose={() => setIsOpen(false)}
         onOpen={() => setIsOpen(true)}
       />
-    </Layout>
+    </>
   );
 };
