+++ src/components/parts/Footer/index.stories.tsx
@@ -1,6 +1,5 @@
-import { Layout } from "../Layout";
 import { Footer } from ".";
-import type { Meta, StoryObj } from "@storybook/nextjs";
+import type { Meta, StoryObj } from "@storybook/nextjs-vite";
 
 const meta: Meta<typeof Footer> = {
   component: Footer,
@@ -9,15 +8,6 @@ const meta: Meta<typeof Footer> = {
 export default meta;
 type Story = StoryObj<typeof Footer>;
 
-const Component: Story["render"] = () => {
-  return (
-    <Layout>
-      <Footer />
-    </Layout>
-  );
-};
-
 export const Default: Story = {
   args: {},
-  render: Component,
 };
